rename upvote to upvotePopover
