--- src/components/upvote/index.js
@@ -1,4 +1,0 @@
-import Upvote from './container/upvoteContainer';
-
-export { Upvote };
-export default Upvote;
