+++ src/components/upvotePopover/children/upvoteStyles.js
