+++ src/components/upvotePopover/index.js
@@ -1,0 +1,4 @@
+import UpvotePopover from './container/upvotePopover';
+
+export { UpvotePopover };
+export default UpvotePopover;
